+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,25 @@
 # axeom-docs
 
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @axeom/express@0.2.1
+  - @axeom/core@0.2.1
+  - @axeom/framework@0.2.1
+  - @axeom/web@0.2.1
+  - @axeom/client@0.2.1
+  - @axeom/auth@0.2.1
+  - @axeom/compression@0.2.1
+  - @axeom/cors@0.2.1
+  - @axeom/rate-limit@0.2.1
+  - @axeom/security@0.2.1
+  - @axeom/static@0.2.1
+  - @axeom/swagger@0.2.1
+  - @axeom/upload@0.2.1
+  - @axeom/ws@0.2.1
+
 ## 0.2.0
 
 ### Patch Changes