+++ examples/cloudflare-workers/CHANGELOG.md
@@ -1,5 +1,12 @@
 # cloudflare-workers-example
 
+## 0.2.1
+
+### Patch Changes
+
+- @axeom/framework@0.2.1
+- @axeom/swagger@0.2.1
+
 ## 0.2.0
 
 ### Patch Changes