+++ packages/plugins/upload/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @axeom/upload
 
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @axeom/core@0.2.1
+
 ## 0.2.0
 
 ### Minor Changes